Create Unit test for controllers and models (#6)
* Create Rspec configuration

Add a gem "Rspec rails" in gemfile and create some files for the project
test.

* Add test for controllers and views

Configurate, add files and test controllers and models for lists index
and validations. Add a test for the welcome controller.

* Add the authenticate user before any action

* Add shoulda matchers gem for the controller tests

* Create a first view for the root

Generate a controller for the welcome view to manage the user log
in, sign up and log out, also modify the routes to make welcome our
first view.

* Fix user, list and task models

Add a dependent destroy to user and list models and some validations
for the tasks.

* Add messages for errors and switch user

Edit form of tasks and add messages for errors and welcome condition
to sign in and log out.

* Create Unit test for controllers and models

Validate the creation of a user with the correct parameters, test
the relation of the databases and the syntax of the parameters in
lists and tasks.