Develop (#21)
* Create list CRUD with a title

Controller, views and model is generate for the user with the
scaffold command.

* Create Task CRUD and relate to List CRUD

Modify some list files to show the tasks in the list which belong, and
Task CRUD generate with the model(database and test files), controller
(control the actions of Create, Read, Update and Delete) and view and
asset files.

Resolves: #2,#3

* Create relation user with lists

Modify user and list models to create the relation and make a new
migration to generate a foreign key to lists database.

* Create Unit test for controllers and models (#6)

* Create Rspec configuration

Add a gem "Rspec rails" in gemfile and create some files for the project
test.

* Add test for controllers and views

Configurate, add files and test controllers and models for lists index
and validations. Add a test for the welcome controller.

* Add the authenticate user before any action

* Add shoulda matchers gem for the controller tests

* Create a first view for the root

Generate a controller for the welcome view to manage the user log
in, sign up and log out, also modify the routes to make welcome our
first view.

* Fix user, list and task models

Add a dependent destroy to user and list models and some validations
for the tasks.

* Add messages for errors and switch user

Edit form of tasks and add messages for errors and welcome condition
to sign in and log out.

* Create Unit test for controllers and models

Validate the creation of a user with the correct parameters, test
the relation of the databases and the syntax of the parameters in
lists and tasks.

* Create profile of a user

Add a controller to show the current user profile with name and description
For that was necessary to create a migration to add name and description to
the user database. Finally, generate devise views to  manipulate the sign in
and log out styles.

* Add routes for every view

Link welcome, task, list and user views for users to have control of
each one and Add users route for the controller.

* Fix update users

Change controller application to allow "name, description and password"
fields in the sign in, sign up and account update.

* Clean code of the rails app

Make the code more understandable, with a good structure and easy
to read for another developer.

* Export tables to PDF and CSV

Modify models to add the csv function that generate csv files and
add a respond_to in the controller to accept the .csv files.

Install a wicked gem to convert html to pdf files and add to the
controller a format .pdf to render the template.

* Configure the sidekiq, mailers and jobs to send emails (#12)

* Configure mailer to create mails in local

Create method and views to send emails to see them in letter opener

* Configure job and sidekiq-schedule

Create a job to send the mailer in a time specified by the sidekiq
schedule that appears in the sidekiq control panel.

* Create the design for the sign in, sign up and forget password views

Add the bootstrap gem to modify the views with classes and add some
resources to make sense of the purpose of the application.

- Fix a bug of the sign in with the jquery gem.

Resolves: #9

* Create the welcome design

This customize the home page with a good bootstrap design.

* Create the task and list views

This commit change the styles of the show, edit and new of the tasks
and lists views using bootstrap and images to make them better.

* Add the design of the profile views

This makes the profile show and edit views better with bootstrap.

* Modify the PDF and CSV views

This commit make a better design for the pdf files.

- Fix an error in the error messages of the tasks views

Resolves: #18

* Create integration test with capybara

This commit use factory bot and capybara to make a great flow of the
views and the elements or content that they have.

- Fix some placeholders and add ids to assign them in the capybara test
- Add to the users controller the factory configuration
- Add in the gemfile factory bot and capybara

Resolves: #16

* Modify all the code to be more clean

This commit creates a better code to be more readable and structured.

* Add and modify profile avatar

This commit allows to the user to update his profile with
active storage.

- Fix sidekiq schedule mailers
- Add a validation of the user profile image

Resolves: #15